+++ Source/KrzyweKarty/Zakon/Zakon_Paladyn.h
@@ -17,7 +17,7 @@ class KRZYWEKARTY_API AZakon_Paladyn : public AKKCharacter, public IAreaModifier
 	
 public:
 
-	virtual void PerformAbility_Implementation(int32 Index) override;
+	virtual void PerformAbility_Implementation(uint8 Index) override;
 
-	TArray<FDirection> GetAffectedTiles_Implementation(int32 Index) override;
+	TArray<FDirection> GetAffectedTiles_Implementation(uint8 Index) override;
 };
